+++ src/Cassandra/DataStax/Cloud/CustomCACertificateValidator.cs
@@ -15,8 +15,10 @@
 //
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net.Security;
+using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 
 namespace Cassandra.DataStax.Cloud
@@ -26,9 +28,54 @@ namespace Cassandra.DataStax.Cloud
     /// </summary>
     internal class CustomCaCertificateValidator : ICertificateValidator
     {
+        private const string SubjectAlternateNameOid = "2.5.29.17"; // Oid for the SAN extension
+
         private static readonly Logger Logger = new Logger(typeof(CustomCaCertificateValidator));
+
+        private static readonly string SanPlatformId;
+        private static readonly string SanSeparator;
+
         private readonly X509Certificate2 _trustedRootCertificateAuthority;
         private readonly string _hostname;
+
+        static CustomCaCertificateValidator()
+        {
+            // use a well known example SAN extension to extract the platform identifier (since it is affected by platform and locale/culture)
+
+            const string wellKnownSanExtension = @"MBuCC2V4YW1wbGUuY29tggxmYWtlLXN1YmplY3Q=";
+            const string firstWellKnownDomainName = "example.com";
+            const string secondWellKnownDomainName = "fake-subject";
+
+            var formattedExtString = new X509Extension(SubjectAlternateNameOid, Convert.FromBase64String(wellKnownSanExtension), true).Format(false);
+
+            // Windows identifier is affected by locale/culture
+            // Example well known SAN extension has the following format:
+            // Windows: "DNS Name=example.com, DNS Name=fake-subject"
+            // Linux: "DNS:example.com, DNS:fake-subject"
+            //
+            // Parse it as the following:
+            // <platform-id><domain-name><separator><platform-id><domain-name>
+            // e.g. for Windows with EN culture:
+            //      platform-id -> "DNS Name="
+            //      first-domain-name -> "example.com"
+            //      second-domain-name -> "fake-subject"
+            //      separator   -> ", "
+
+            // "example.com"
+            var firstDomainNameIndex = formattedExtString.IndexOf(firstWellKnownDomainName, StringComparison.Ordinal);
+
+            // "fake-subject"
+            var secondDomainNameIndex = formattedExtString.IndexOf(secondWellKnownDomainName, StringComparison.Ordinal);
+
+            // "DNS Name="
+            SanPlatformId = formattedExtString.Substring(0, firstDomainNameIndex);
+
+            // ", "
+            var separatorIndex = firstDomainNameIndex + firstWellKnownDomainName.Length;
+            var lengthUntilSeparator = firstDomainNameIndex + firstWellKnownDomainName.Length;
+            var separatorLength = secondDomainNameIndex - SanPlatformId.Length - lengthUntilSeparator;
+            SanSeparator = formattedExtString.Substring(separatorIndex, separatorLength);
+        }
 
         public CustomCaCertificateValidator(X509Certificate2 trustedRootCertificateAuthority, string hostname)
         {
@@ -58,28 +105,25 @@ namespace Cassandra.DataStax.Cloud
             {
                 GetOrCreateCert2(ref cert2, cert);
                 var cn = cert2.GetNameInfo(X509NameType.SimpleName, false);
+                var subjectAlternativeNames = GetSubjectAlternativeNames(cert2).ToList();
+                var names = new List<string> { cn }.Concat(subjectAlternativeNames);
                 var validName = false;
-                if (cn == null || _hostname == null || !cn.StartsWith("*."))
+
+                foreach (var name in names)
                 {
-                    if (cn == _hostname)
+                    validName = ValidateName(name);
+                    if (validName)
                     {
-                        validName = true;
-                    }
-                } else if (cn.StartsWith("*."))
-                {
-                    cn = cn.Remove(0, 1);
-                    if (_hostname.EndsWith(cn))
-                    {
-                        validName = true;
+                        break;
                     }
                 }
 
                 if (!validName)
                 {
                     CustomCaCertificateValidator.Logger.Error(
-                        "Failed to validate the server certificate's CN. Expected {0} but found {1}.",
+                        "Failed to validate the server certificate's CN. Expected {0} but found CN={1} and SANs={2}.",
                         _hostname,
-                        cn ?? "null");
+                        cn ?? "null", string.Join(",", subjectAlternativeNames));
                 }
 
                 valid = validName;
@@ -143,6 +187,46 @@ namespace Cassandra.DataStax.Cloud
 
             DisposeCert2(cert2);
             return valid;
+        }
+
+        private bool ValidateName(string name)
+        {
+            if (name == null || _hostname == null || !name.StartsWith("*."))
+            {
+                if (name?.ToLowerInvariant() == _hostname?.ToLowerInvariant())
+                {
+                    return true;
+                }
+            }
+            else if (name.StartsWith("*."))
+            {
+                name = name.Remove(0, 1);
+                if (_hostname.EndsWith(name, StringComparison.InvariantCultureIgnoreCase))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        /// Check the static constructor inline comments for an explanation about this
+        /// </summary>
+        private IEnumerable<string> GetSubjectAlternativeNames(X509Certificate2 cert)
+        {
+            var sanStrings = cert.Extensions
+                .Cast<X509Extension>()
+                .Where(ext => ext.Oid.Value == SubjectAlternateNameOid)
+                .Select(ext => new AsnEncodedData(ext.Oid, ext.RawData).Format(false)).ToList();
+
+            var splitSanStrings = sanStrings.SelectMany(s => s.Split(new[] { SanSeparator }, StringSplitOptions.RemoveEmptyEntries));
+
+            // remove the platform identifier (i.e. "DNS Name=") from the strings to get the domain names
+            return splitSanStrings
+                .Where(s => s.StartsWith(SanPlatformId) && s.Length > SanPlatformId.Length)
+                .Select(s => s.Substring(SanPlatformId.Length));
+
         }
 
         private void GetOrCreateCert2(ref X509Certificate2 cert2, X509Certificate cert)