+++ src/Cassandra/Serialization/Serializer.cs
@@ -21,11 +21,13 @@ namespace Cassandra.Serialization
     internal class Serializer : ISerializer
     {
         private readonly IGenericSerializer _serializer;
-        
-        public Serializer(ProtocolVersion version, IGenericSerializer serializer)
+        private readonly IColumnEncryptionPolicy _columnEncryptionPolicy;
+
+        public Serializer(ProtocolVersion version, IGenericSerializer serializer, IColumnEncryptionPolicy columnEncryptionPolicy)
         {
             ProtocolVersion = version;
             _serializer = serializer;
+            _columnEncryptionPolicy = columnEncryptionPolicy;
         }
 
         public ProtocolVersion ProtocolVersion { get; }
@@ -35,14 +37,106 @@ namespace Cassandra.Serialization
             return _serializer.Deserialize(ProtocolVersion, buffer, offset, length, typeCode, typeInfo);
         }
 
+        public object DeserializeAndDecrypt(string ks, string table, string column, byte[] buffer, int offset, int length, ColumnTypeCode typeCode, IColumnInfo typeInfo)
+        {
+            var columnEncryptionMetadata = _columnEncryptionPolicy.GetColumnEncryptionMetadata(ks, table, column);
+            if (columnEncryptionMetadata != null)
+            {
+                var encryptedData = _serializer.Deserialize(ProtocolVersion, buffer, offset, length, typeCode, typeInfo);
+                if (encryptedData == null)
+                {
+                    throw new DriverInternalError("deserialization of encrypted data returned null");
+                }
+
+                if (!(encryptedData is byte[] encryptedDataBuf))
+                {
+                    throw new ColumnEncryptionInvalidTypeError(column, typeCode, encryptedData);
+                }
+
+                var decryptedDataBuf = _columnEncryptionPolicy.Decrypt(columnEncryptionMetadata.Value.Key, encryptedDataBuf);
+                if (decryptedDataBuf == null)
+                {
+                    return null;
+                }
+                return _serializer.Deserialize(ProtocolVersion, decryptedDataBuf, 0, decryptedDataBuf.Length, columnEncryptionMetadata.Value.TypeCode, columnEncryptionMetadata.Value.TypeInfo);
+            }
+            return _serializer.Deserialize(ProtocolVersion, buffer, offset, length, typeCode, typeInfo);
+        }
+
         public byte[] Serialize(object value)
         {
             return _serializer.Serialize(ProtocolVersion, value);
         }
 
+        public byte[] SerializeAndEncrypt(string defaultKs, RowSetMetadata metadata, int colIdx, object[] values, int valueIdx)
+        {
+            if (values == null)
+            {
+                throw new InvalidOperationException("could not serialize value because values array is null");
+            }
+
+            if (valueIdx >= values.Length)
+            {
+                throw new InvalidOperationException($"could not serialize value because value index {valueIdx} is out of bounds {values.Length}");
+            }
+
+            var value = values[valueIdx];
+
+            if (!IsEncryptionEnabled)
+            {
+                return Serialize(value);
+            }
+
+            byte[] serialized;
+            if (value is EncryptedValue markedValue)
+            {
+                serialized = _serializer.Serialize(ProtocolVersion, markedValue.Value);
+                serialized = _columnEncryptionPolicy.Encrypt(markedValue.Key, serialized);
+                return _serializer.Serialize(ProtocolVersion, serialized);
+            }
+
+            serialized = _serializer.Serialize(ProtocolVersion, value);
+
+            if (metadata == null)
+            {
+                // probably simple statement
+                return serialized;
+            }
+
+            if (metadata.Columns == null)
+            {
+                throw new InvalidOperationException("column metadata is null");
+            }
+
+            if (colIdx >= metadata.Columns.Length)
+            {
+                throw new InvalidOperationException($"column index is {colIdx}, columns length is {metadata.Columns.Length}");
+            }
+
+            var colMetadata = metadata.Columns[colIdx];
+            var encryptionMetadata = _columnEncryptionPolicy.GetColumnEncryptionMetadata(
+                colMetadata.Keyspace ?? metadata.Keyspace ?? defaultKs, colMetadata.Table, colMetadata.Name);
+            if (encryptionMetadata == null)
+            {
+                return serialized;
+            }
+
+            serialized = _columnEncryptionPolicy.Encrypt(encryptionMetadata.Value.Key, serialized);
+            return _serializer.Serialize(ProtocolVersion, serialized);
+
+        }
+
         public ISerializer CloneWithProtocolVersion(ProtocolVersion version)
         {
-            return new Serializer(version, _serializer);
+            return new Serializer(version, _serializer, _columnEncryptionPolicy);
+        }
+
+        public bool IsEncryptionEnabled => _columnEncryptionPolicy != null;
+
+        public Tuple<bool, ColumnTypeCode> IsAssignableFromEncrypted(string ks, string table, string column, ColumnTypeCode columnTypeCode, object value)
+        {
+            var colData = _columnEncryptionPolicy.GetColumnEncryptionMetadata(ks, table, column);
+            return new Tuple<bool, ColumnTypeCode>(IsAssignableFrom(colData?.TypeCode ?? columnTypeCode, value), colData?.TypeCode ?? columnTypeCode);
         }
 
         public object Deserialize(ProtocolVersion version, byte[] buffer, int offset, int length, ColumnTypeCode typeCode, IColumnInfo typeInfo)
@@ -75,9 +169,9 @@ namespace Cassandra.Serialization
             return _serializer.GetCqlType(type, out typeInfo);
         }
 
-        public bool IsAssignableFrom(CqlColumn column, object value)
+        public bool IsAssignableFrom(ColumnTypeCode columnTypeCode, object value)
         {
-            return _serializer.IsAssignableFrom(column, value);
+            return _serializer.IsAssignableFrom(columnTypeCode, value);
         }
 
         public UdtMap GetUdtMapByName(string name)